% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_acf.R
\name{dts_acf_plot}
\alias{dts_acf_plot}
\title{Plot Autocorrelation and Partial Autocorrelation Functions (ACF/PACF)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_acf_plot(
  data,
  .partial = TRUE,
  .mts = FALSE,
  .interactions = FALSE,
  .select = NULL
)
}
\arguments{
\item{data}{A time series object of class \code{ts} or \code{mts}, or a data frame with a "date" column and
one or more numeric columns representing the time series.}

\item{.partial}{Logical. Should the PACF be plotted as well? Default is TRUE.}

\item{.mts}{Logical. Is the input data a multivariate time series (\code{mts})? Default is FALSE.}

\item{.interactions}{Logical. Should interactions between series be considered (not yet implemented)?
Default is FALSE.}

\item{.select}{Character vector. Names of columns to select (if \code{data} is multivariate). Default is NULL (all columns).}
}
\value{
A ggplot object displaying the ACF and/or PACF.
}
\description{
This function creates a ggplot2-based plot of the Autocorrelation Function (ACF) and, optionally,
the Partial Autocorrelation Function (PACF) for a time series.
}
\examples{
\dontrun{
  data(AirPassengers)
  dts_acf_plot(AirPassengers)

  data(economics, package = "ggplot2")
  dts_acf_plot(economics, .select = c("psavert", "uempmed"))
}
}
