% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_breakspoints.R
\name{dts_breakpoints}
\alias{dts_breakpoints}
\title{Detect and visualize breakpoints in time series data
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_breakpoints(.ts, .n = NULL, .bps = c("s", "a", "n"), ...)
}
\arguments{
\item{.ts}{A time series object of class \code{ts}.}

\item{.n}{The maximum number of breakpoints to consider. If NULL (default), a maximum of 3 breakpoints
or a third of the number of years in the time series (whichever is smaller) is used.}

\item{.bps}{The type of breakpoint information to return:
\itemize{
\item "s": A summary of the breakpoint analysis (default).
\item "a": The raw breakpoint object from \code{strucchange::breakpoints()}.
\item "n": No breakpoint information, only plots are returned.
}}

\item{...}{Additional arguments passed to \code{strucchange::breakpoints()}.}
}
\value{
A list containing:
\itemize{
\item Plots visualizing the detected breakpoints for each considered number of breaks.
\item Optionally, breakpoint information (\code{summary} or the \code{breakpoints} object) based on the \code{.bps} argument.
}
}
\description{
This function uses the \code{strucchange} package to detect breakpoints in time series data
and generates plots to visualize the detected breakpoints.
}
\examples{
\dontrun{
  data(AirPassengers)
  bp_analysis <- dts_breakpoints(AirPassengers, .n = 4)
  bp_analysis$bps
  bp_analysis[[1]]

  # Return only plots
  bp_plots <- dts_breakpoints(AirPassengers, .n = 2, .bps = "n")
}
}
