% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_concat.R
\name{dts_concat}
\alias{dts_concat}
\title{Concatenate time series objects
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_concat(...)
}
\arguments{
\item{...}{Time series objects (\code{ts} or \code{mts}) to be concatenated.}
}
\value{
A concatenated time series object.
}
\description{
This function concatenates multiple time series objects into a single time series object,
ensuring that the concatenated series is contiguous.
}
\examples{
\dontrun{
  data(AirPassengers)
  ts1 <- window(AirPassengers, end = c(1955, 12))
  ts2 <- window(AirPassengers, start = c(1956, 1))
  concatenated_ts <- dts_concat(ts1, ts2)
  concatenated_ts
}
}
