% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_ts.R
\name{dts_df_to_ts}
\alias{dts_df_to_ts}
\title{Convert a Data Frame to a Time Series Object
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_df_to_ts(data, .freq = NULL, .date = "date", end = NULL)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{.freq}{The frequency of the time series (e.g., 12 for monthly, 4 for quarterly). If NULL (default),
the frequency is automatically estimated based on the date column.}

\item{.date}{The name of the date column in the data frame (default is "date").}

\item{end}{The end date of the time series. If NULL (default), the maximum date in the data is used.}
}
\value{
A time series object of class \code{ts} or \code{mts}.
}
\description{
This function converts a data frame with a date column into a time series object (\code{ts} or \code{mts}).
}
\examples{
\dontrun{
  data(fdeaths, package = "datasets")
  ts_data <- data.frame(date = as.Date(time(fdeaths)), y = fdeaths)
  ts_obj <- dts_df_to_ts(ts_data, .freq = 12)
  ts_obj
}
}
