% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_diff.R
\name{dts_diff}
\alias{dts_diff}
\title{Calculate the difference of a time series
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_diff(.ts, .n = 1, .subset = NULL, .na.omit = TRUE, ...)
}
\arguments{
\item{.ts}{A time series object of class \code{ts} or \code{mts}.}

\item{.n}{An integer specifying the number of lags to use for calculating the difference (default is 1).}

\item{.subset}{A character vector specifying the columns to be differentiated (default is all columns).}

\item{.na.omit}{Logical. Should missing values be omitted after differencing? Default is TRUE.}

\item{...}{Additional arguments passed to other functions (currently unused).}
}
\value{
A time series object of the same class as the input, with the differences calculated.
}
\description{
This function calculates the difference of a time series, preserving the date and
\code{.name} attribute (if present). It provides more control over the differencing process
compared to the base \code{diff()} function.
}
\examples{
\dontrun{
  data(AirPassengers)
  diff_ts <- dts_diff(AirPassengers)
  diff_ts

  # Difference with 2 lags
  diff_ts_2 <- dts_diff(AirPassengers, .n = 2)
  diff_ts_2
}
}
