% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_dummy.R
\name{dts_dummy}
\alias{dts_dummy}
\title{Create time series dummy variables
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_dummy(
  dates,
  start = NULL,
  end = NULL,
  frequency = NULL,
  .name = NULL,
  ts = NULL,
  method = "==",
  ...
)
}
\arguments{
\item{dates}{A vector, data frame, or list of dates to set as 1 in the dummy variable.}

\item{start}{The start date of the time series (required if \code{ts} is not provided).}

\item{end}{The end date of the time series (required if \code{ts} is not provided).}

\item{frequency}{The frequency of the time series (required if \code{ts} is not provided).}

\item{.name}{An optional name for the output time series. Default is NULL.}

\item{ts}{A time series object (\code{ts} or \code{mts}). If provided, \code{start}, \code{end}, and \code{frequency} are ignored.}

\item{method}{The comparison operator to use when matching dates: "<", "<=", "==", ">=", ">". Default is "==".}

\item{...}{Additional arguments to be passed to \code{ts}.}
}
\value{
A time series object (\code{ts} or \code{mts}) representing the dummy variable(s).
}
\description{
This function generates time series dummy variables based on provided dates, a time series object,
or start/end/frequency specifications.
}
\examples{
\dontrun{
  # Using dates and start/end/frequency
  tsd1 <- dts_dummy(dates = 2005, start = 2000, end = 2011, frequency = 1)
  tsd1

  tsd2 <- dts_dummy(dates = c(2005, 3), start = 2000, end = c(2011, 4), frequency = 4)
  tsd2

  tsd3 <- dts_dummy(dates = data.frame(c(2005, 2006), c(8, 11)),
                   start = 2000, end = c(2011, 12), frequency = 12)
  tsd3

  tsd4 <- dts_dummy(dates = list("2005M07" = c(2005, 7), "2006M11" = c(2006, 11)),
                   start = 2000, end = 2011, frequency = 12)

  tsd3

  # Using a time series object
  data(AirPassengers)
  tsd5 <- dts_dummy(dates = c(1955, 6), ts = AirPassengers)
  tsd5
}
}
