% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_extend.R
\name{dts_extend}
\alias{dts_extend}
\title{Extend a time series to a specified end date
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_extend(.ts, .end, .default = 0, .sub = TRUE)
}
\arguments{
\item{.ts}{The time series to be extended.}

\item{.end}{The desired end date for the extended time series.}

\item{.default}{Numeric. The default value to use for filling missing values in the extended portion. Default is 0.}

\item{.sub}{Logical. Should the original time series be subtracted from the extended time series,
leaving only the extended part? Default is TRUE.}
}
\value{
A time series object containing only the extended portion of the original time series.
}
\description{
This function extends a time series to a given end date, filling missing values with a default value.
It returns only the extended portion of the time series, excluding the original data.
}
\examples{
\dontrun{
  data(AirPassengers)
  dts_extend(AirPassengers, .end = c(1965, 12))
  dts_extend(AirPassengers, .end = c(1965, 12), .default = 100)
  dts_extend(AirPassengers, .end = c(1965, 12), .sub = FALSE)
}
}
