% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_merge.R
\name{dts_merge}
\alias{dts_merge}
\title{Merge multiple time series
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_merge(..., .by = "date", .type = "full")
}
\arguments{
\item{...}{Time series objects of class \code{ts} or \code{mts}.}

\item{.by}{The column to use for merging. Defaults to "date".}

\item{.type}{The type of join to perform. Can be one of "full", "inner", "left",
or "right". See \code{dplyr::full_join} for more information. Defaults to "full".}
}
\value{
A merged time series object of class \code{mts}.
}
\description{
This function merges multiple time series objects based on their dates. It handles
both univariate (\code{ts}) and multivariate (\code{mts}) time series. If the input time
series are not of class \code{mts}, the column names will be used to label the
respective series in the merged data.
}
\examples{
\dontrun{
  ts1 <- ts(rnorm(100), start = c(2010, 1), frequency = 12)
  ts2 <- ts(rnorm(50), start = c(2010, 1), frequency = 12)
  ts3 <- ts(rnorm(120), start = c(2010, 1), frequency = 12)
  merged_ts <- dts_merge(ts1, ts2, ts3)
  merged_ts
}
}
