% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_name.R
\name{dts_name}
\alias{dts_name}
\title{Set a name to a time series
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_name(.ts, .name)
}
\arguments{
\item{.ts}{A time series object of class \code{ts} or \code{mts}.}

\item{.name}{A character string specifying the name to assign to the time series.}
}
\value{
The time series object with the name set in the \code{.name} attribute.
}
\description{
This function sets a custom name to a time series object, storing the name in
the \code{.name} attribute.
}
\examples{
\dontrun{
  data(AirPassengers)
  ts_air <- dts_name(AirPassengers, "AirPassengersData")
  attr(ts_air, ".name")
}
}
