% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_resize.R
\name{dts_resize}
\alias{dts_resize}
\title{Resize a time series to match another
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_resize(.ts, .resize, .start = TRUE, .default = 0)
}
\arguments{
\item{.ts}{The time series to be resized.}

\item{.resize}{The reference time series for resizing.}

\item{.start}{Logical. Should the start date of the resized time series match the start date of the
reference time series? If FALSE, the original start date is kept. Default is TRUE.}

\item{.default}{Numeric. The default value to use for filling missing values. Default is 0.}
}
\value{
A resized time series object with the same frequency as the reference time series.
}
\description{
This function resizes a time series to match the start and end dates of another time series.
It handles missing values by filling them with a specified default value.
}
\examples{
\dontrun{
  data(AirPassengers)

  # Resize to a shorter period
  short_ts <- window(AirPassengers, start = c(1955, 1), end = c(1958, 12))
  resized_ts <- dts_resize(AirPassengers, short_ts)
  resized_ts

  # Resize to a longer period, keeping the original start date
  long_ts <- window(AirPassengers, start = c(1949, 1), end = c(1962, 12))
  resized_ts <- dts_resize(AirPassengers, long_ts, .start = FALSE, .default = 0)
  resized_ts
}
}
