% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_seas.R
\name{dts_seas}
\alias{dts_seas}
\title{Create seasonal dummy variables for a time series
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_seas(.ts, .periods = NULL, .rm = FALSE)
}
\arguments{
\item{.ts}{A time series object of class \code{ts} or \code{mts}.}

\item{.periods}{An optional vector of periods (quarters, months, or days) to create dummy variables for.
If NULL (default), dummy variables are created for all periods except the last one.}

\item{.rm}{Logical. Should existing data columns be removed from the output? Default is FALSE.}
}
\value{
A time series object with added seasonal dummy variables.
}
\description{
This function generates seasonal dummy variables for a time series object, representing
quarters, months, or days, depending on the frequency of the time series.
}
\examples{
\dontrun{
  data(AirPassengers)
  ts_with_dummies <- dts_seas(AirPassengers)
  ts_with_dummies

  # Create dummies for specific months
  ts_with_dummies <- dts_seas(AirPassengers, .periods = c(1, 7, 12))
  ts_with_dummies
}
}
