% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_stationarity_test.R
\name{dts_stationarity_test}
\alias{dts_stationarity_test}
\title{Test stationarity of time series using Augmented Dickey-Fuller Test
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_stationarity_test(data, .select = NULL)
}
\arguments{
\item{data}{A time series object of class \code{ts} or \code{mts}.}

\item{.select}{An optional character vector of column names to select for testing
(used only if \code{data} is a \code{mts} object).}
}
\value{
A list of ADF test results (for multivariate input) or a single ADF test result (for univariate input).
}
\description{
This function performs the Augmented Dickey-Fuller (ADF) test to check for stationarity in
a time series. It handles both univariate and multivariate time series.
}
\examples{
\dontrun{
  # Univariate example
  data(AirPassengers)
  ts_stationarity_test(AirPassengers)

  # Multivariate example
  ts1 <- ts(rnorm(100), start = c(2010, 1), frequency = 12)
  ts2 <- ts(rnorm(100), start = c(2010, 1), frequency = 12)
  data <- dts_merge(ts1, ts2)
  dts_stationarity_test(data)
  dts_stationarity_test(data, .select = "ts1")
}
}
