% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_sub.R
\name{dts_sub}
\alias{dts_sub}
\title{Subtract two time series
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_sub(ts1, ts2)
}
\arguments{
\item{ts1}{The time series to subtract from (minuend).}

\item{ts2}{The time series to subtract (subtrahend).}
}
\value{
A time series object representing the difference between \code{ts1} and \code{ts2}.
}
\description{
This function subtracts one time series (\code{ts2}) from another (\code{ts1}).
}
\examples{
\dontrun{
  ts1 <- ts(rnorm(100), start = c(2010, 1), frequency = 12)
  ts2 <- ts(rnorm(50), start = c(2010, 1), frequency = 12)
  diff_ts <- dts_sub(ts1, ts2)
  diff_ts
}
}
