% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{dts_test_split}
\alias{dts_test_split}
\title{Get test set from a sataset split
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_test_split(tts)
}
\arguments{
\item{tts}{A list containing the train, test, and optionally valid sets (output of \code{train_test_split}).}
}
\value{
The test set.
}
\description{
Get test set from a sataset split
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
  data(AirPassengers)
  split_data <- dts_train_test_split(AirPassengers, prop = 0.8)
  test_set <- dts_test_split(split_data)
  test_set
}
}
