% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{dts_train_test_split}
\alias{dts_train_test_split}
\title{Split time series data into train and test Sets
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_train_test_split(data, prop = 3/4, val_split = FALSE)
}
\arguments{
\item{data}{A time series object of class \code{ts} or \code{mts}.}

\item{prop}{The proportion or number of observations for the testing set:
\itemize{
\item If \code{prop < 1}, it's treated as a proportion of the total observations.
\item If \code{prop >= 1}, it's treated as the exact number of observations.
}}

\item{val_split}{Logical. Should a validation set be created from the test set? Default is FALSE.}
}
\value{
A list containing the train and test sets, and optionally a validation set:
\itemize{
\item \code{train}: The training set.
\item \code{test}: The test set.
\item \code{valid}: The validation set (if \code{val_split = TRUE}).
}
}
\description{
This function splits time series data into train and test sets based on a specified proportion
or number of observations. It also provides an option to create a validation set.
}
\examples{
\dontrun{
  data(AirPassengers)
  split_data <- dts_train_test_split(AirPassengers, prop = 0.2)
  dts_train_split(split_data)
  dts_test_split(split_data)

  # With a validation set
  split_data <- dts_train_test_split(AirPassengers, prop = 0.3, val_split = TRUE)
  dts_train_split(split_data)
  dts_valid_split(split_data)
  dts_test_split(split_data)
}
}
