% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_to_df.R
\name{dts_ts_to_df}
\alias{dts_ts_to_df}
\title{Convert a time series object to a data.frame (Monthly/Quarterly)
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_ts_to_df(ts, as_date = TRUE, .name = NULL, ...)
}
\arguments{
\item{ts}{A time series object of class \code{ts} or \code{mts} with a monthly or quarterly frequency.}

\item{as_date}{Logical. Should a date column be included in the output? Default is TRUE.}

\item{.name}{A character string specifying the name of the value column.
If NULL (default), the name will be taken from the \code{ts} object if available.}

\item{...}{Additional arguments passed to other functions (currently unused).}
}
\value{
A data frame with a \code{date} column (if \code{as_date = TRUE}) and a
column containing the time series values.
}
\description{
This function converts a time series object (\code{ts}) into a data frame, optionally
including a date column. It currently supports \strong{monthly and quarterly} time series.
}
