% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{dts_valid_split}
\alias{dts_valid_split}
\title{Get validation set from a dataset split
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_valid_split(tts)
}
\arguments{
\item{tts}{A list containing the train, test, and optionally valid sets (output of \code{train_test_split}).}
}
\value{
The validation set (NULL if not present).
}
\description{
Get validation set from a dataset split
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
  data(AirPassengers)
  split_data <- dts_train_test_split(AirPassengers, prop = 0.8, val_split = TRUE)
  valid_set <- dts_valid_split(split_data)
  valid_set
}
}
