% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_to_date.R
\name{dts_vars_to_date}
\alias{dts_vars_to_date}
\title{Multiple variables to unique date variable
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_vars_to_date(
  tbl,
  year = NULL,
  quarter = NULL,
  month = NULL,
  day = NULL,
  date = NULL,
  drop_vars = TRUE,
  clean_names = FALSE,
  date_format = "\%d-\%m-\%y",
  origin = "1900-01-01",
  .round = c("end", "middle", "start")
)
}
\arguments{
\item{tbl}{A data frame or tibble.}

\item{year}{The position or name of the year variable.}

\item{quarter}{The position or name of the quarter variable.}

\item{month}{The position or name of the month variable.}

\item{day}{The position or name of the day variable.}

\item{date}{A variable name or position containing a date-like format.}

\item{drop_vars}{Logical, indicating whether to drop the original date component variables. Default is TRUE.}

\item{clean_names}{Logical, indicating whether to clean variable names using \code{janitor::make_clean_names()}. Default is FALSE.}

\item{date_format}{The date format of the \code{date} variable (if provided). Default is "\%d-\%m-\%y".}

\item{origin}{The base date for converting numeric dates. Default is "1900-01-01".}

\item{.round}{Specifies how to round the resulting date: "end", "middle", or "start" of the period. Default is "end".}
}
\value{
A tibble with the computed date variable.
}
\description{
This function converts multiple date-related variables (year, quarter, month, day)
into a single date variable within a data frame. It handles various date formats
and frequencies, offering flexibility in how dates are represented in the input data.
}
\examples{
\dontrun{
  tbl <- data.frame(
    year = rep("2021", 12),
    month = month.name,
    day = sample(1:3, 12, TRUE),
    value = sample(100:1000, 12, TRUE)
  )

  dts_vars_to_date(tbl, year = 1, month = 2, day = 3)

  # With quarters
  tbl <- data.frame(
    year = rep("2021", 12),
    quarter = sample(c("Q1", "Q2", "Q3", "Q4"), 12, TRUE),
    value = sample(100:1000, 12, TRUE)
  )

  dts_vars_to_date(tbl, year = 1, quarter = 2)
}
}
