% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_zero.R
\name{dts_zero}
\alias{dts_zero}
\title{Create a time series object filled with zeros
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dts_zero(x, .name = NULL, resize = NULL)
}
\arguments{
\item{x}{A time series object of class \code{ts} or \code{mts}.}

\item{.name}{An optional name for the output time series. Default is NULL.}

\item{resize}{An optional time series object to use for resizing the output. Default is NULL (no resizing).}
}
\value{
A time series object filled with zeros, with the same class as the input.
}
\description{
This function creates a time series object filled with zeros, with the same length and frequency
as the input time series. Optionally, you can resize the output time series using the \code{resize} argument.
}
\examples{
\dontrun{
  data(AirPassengers)
  zero_ts <- dts_zero(AirPassengers)
  zero_ts

  # Resize the output
  zero_ts_resized <- dts_zero(AirPassengers, resize = window(AirPassengers, start = c(1955, 1)))
  zero_ts_resized
}
}
